% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameterRotor.R
\name{diameterRotor}
\alias{diameterRotor}
\title{Diameter of the wind turbine propeller}
\usage{
diameterRotor(den, E, Vmean, t, totalEff, unit, acc)
}
\arguments{
\item{den}{Air density |kg/m^3|}

\item{E}{Electricity demand |kWh/year|}

\item{Vmean}{Average wind speed |m/s|}

\item{t}{Time |s|}

\item{totalEff}{Overall efficiency of the wind turbine (gear, rotor, generator) |0-1|}

\item{unit}{Unit of propeller diameter |m|, accepts values TRUE or FALSE}

\item{acc}{Decimal scale}
}
\value{
Default values: 't' = 8760, 'den' = 1.23, 'acc' = 1, 'unit' = TRUE, 'acc' = 1
}
\description{
Calculation of the propeller diameter according to the electricity demand
}
\examples{
diameterRotor(E = 3000, Vmean = 4.5, totalEff = 0.3)
}
\references{
http://www.uwm.edu.pl/kolektory/silownie/obliczenia.html
}
