% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Wind rose}
\usage{
generate(coor, data, title, subTitle)
}
\arguments{
\item{coor}{Geographical coordinates written as uppercase (N, NE, E, SE, S, SW, W, NW)}

\item{data}{Vector frequency of the occurrence of a given wind direction}

\item{title}{Chart title}

\item{subTitle}{Subtitle plot}
}
\value{
The wind rose generates a graph which contains an 8-point scale (N, NE, E, SE, S, SW, W, NW)
}
\description{
Generating a wind rose plot
}
\examples{
generate(coor = c("N","NE","E","SE","S","SW","W","NW"),data = c(2,7,7.5,2.75,12.75,18,1.25,3.25))
}
