% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windEnergy.R
\name{windEnergy}
\alias{windEnergy}
\title{Wind energy}
\usage{
windEnergy(den, Vm, timeV, unit)
}
\arguments{
\item{den}{Air density |kg/m^3|}

\item{Vm}{Momentary speed of the air stream |m/s|}

\item{timeV}{Time of momentary wind speed |s|}

\item{unit}{Wind energy unit, accepts values TRUE or FALSE}
}
\value{
Default values: 'den' = 1.23
}
\description{
Wind energy calculations
}
\examples{
windEnergy(den = 1.23,Vm = 5,timeV = 3600)
}
\references{
http://www.wmae.pl/userfiles/file/Aktualnosci/poradnik_a5.pdf
}
