% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windspeedLVL.R
\name{windspeedLVL}
\alias{windspeedLVL}
\title{Wind speed at different altitude}
\usage{
windspeedLVL(Vo, ho, H, n, unit, acc)
}
\arguments{
\item{Vo}{Wind speed at the measured height |m/s|}

\item{ho}{Height of wind measurement |m|}

\item{H}{Conversion of wind speed to given height |m|}

\item{n}{Roughness of the area |-| 0-1}

\item{unit}{Wind speed unit |m/s|, accepts values TRUE or FALSE}

\item{acc}{Decimal scale}
}
\value{
Default values: 'ho' = 10, 'n' = 0.2, 'acc' = 0, 'unit' = TRUE
}
\description{
Conversion of wind speed at the indicated height
}
\examples{
windspeedLVL(Vo = 12, H = 100, acc=1)
 windspeedLVL(Vo = 12, H = seq(50,100,10), acc=1, unit=FALSE)
}
\seealso{
Value for 'n': "https://github.com/mariusz11363/WindTurbine/blob/master/README.md"
}
