% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindTurbine.R
\docType{package}
\name{WindTurbine-package}
\alias{WindTurbine}
\alias{WindTurbine-package}
\title{WindTurbine: Calculations for a wind turbine}
\description{
The package concerns calculations related to the wind turbine. The library allows to calculate the power of the wind farm, the diameter of the propeller based on the demand for electricity and the conversion of wind speed to different heights. The package also allows you to plot a graph regarding wind directions.
}
\details{
For a complete list of functions with individual help pages, use library(help = "WindTurbine").
}
\author{
Mariusz Kupczyk <mariusz.kupczyk@o2.pl>
}
\keyword{internal}
